feat(groups-filter): move filtering to composable
